pyproject.toml
[project]
name = "frame-rate"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Jeffrey Jose", email = "jeffjosejeff@gmail.com" }
]
dependencies = [
    "typer>=0.12.3",
    "pillow>=10.4.0",
    "tailwind-colors>=1.2.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"frame-rate" = "frame_rate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/frame_rate"]
